% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outer_function.R
\name{outer_function}
\alias{outer_function}
\title{Call innerPkg::inner_function, and print some stuff}
\usage{
outer_function()
}
\description{
Call innerPkg::inner_function, and print some stuff
}
\examples{
outer_function()
}
